pyproject.toml
[tool.poetry]
name = "qortex-test-assignment"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "qortex_test_assignment"}]

[tool.poetry.dependencies]
python = "^3.11"
Django = "^4.1.7"
psycopg2-binary = "^2.9.5"
python-dotenv = "^0.21.1"
drf-yasg = "^1.21.5"
djangorestframework = "^3.14.0"
pytest = "^7.2.1"
factory-boy = "^3.2.1"
pytest-factoryboy = "^2.5.1"
pytest-django = "^4.5.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
